pyproject.toml
[project]
name = "basic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.24.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.0.0",
]
